// src/types/generated/PromptTemplate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface PromptTemplate { id: number | null, name: string, description: string | null, scenario: string, tags: string | null, language: string, isSystem: boolean, createdAt: string, updatedAt: string, }